requirements_images.txt
# Dépendances pour le système d'images Keur Resto
# Pour la présentation de demain

# Images et graphiques
Pillow>=10.0.0

# Requêtes HTTP pour les tests
requests>=2.31.0

# Django et REST Framework (déjà installé mais pour référence)
Django>=5.2.4
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
djangorestframework-simplejwt>=5.2.0

# Notifications push
pyfcm>=1.5.4

# Base de données (SQLite par défaut)
# Pas de dépendance supplémentaire nécessaire

# Pour les tests et développement
pytest>=7.4.0
pytest-django>=4.5.0